Renamed and restarting main program
